# src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: HDUD API Core
  version: 0.7.0
  description: >
    HDUD Core API (v0.7) — Auth + Authors + Memories (versioning, ledger, rollback, timeline).
    Unicode-safe end-to-end (NVARCHAR + mssql NVarChar). All protected endpoints use Bearer JWT.
servers:
  - url: http://localhost:4000
    description: Local Dev
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    MemoryCreateRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 500
          example: "Memória com acentuação correta — versão final"
        content:
          type: string
          example: "ÁÉÍÓÚ — memória máquina coração funcionando corretamente."
      required: [title, content]
    MemoryUpdateRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 500
          example: "Primeira memória fora do domínio (v2)"
        content:
          type: string
          example: "Atualizada — HDUD está operando com versionamento."
      additionalProperties: false
    Memory:
      type: object
      properties:
        memory_id: { type: integer, example: 13 }
        author_id: { type: integer, example: 1 }
        title: { type: string }
        content: { type: string }
        created_at: { type: string, format: date-time }
        version_number: { type: integer, example: 1 }
        is_deleted: { type: boolean, example: false }
      required: [memory_id, author_id, title, content, created_at, version_number, is_deleted]
    MemoryWithMeta:
      allOf:
        - $ref: "#/components/schemas/Memory"
        - type: object
          properties:
            meta:
              type: object
              properties:
                can_edit: { type: boolean, example: true }
                current_version: { type: integer, example: 1 }
    LedgerEvent:
      type: object
      properties:
        ledger_id: { type: string, example: "27" }
        event_type: { type: string, example: MEMORY_UPDATED }
        entity_type: { type: string, example: MEMORY }
        entity_id: { type: string, example: "11" }
        version_number: { type: integer, example: 2 }
        created_at: { type: string, format: date-time }
        created_by: { type: string, example: hdud_api_v0.1 }
        payload:
          type: object
          nullable: true
      required: [ledger_id, event_type, entity_type, entity_id, version_number, created_at, created_by]
    RollbackResponse:
      type: object
      properties:
        message: { type: string, example: "Rollback realizado com sucesso para a versão 1." }
        updated: { $ref: "#/components/schemas/Memory" }
      required: [message, updated]
    TimelineResponse:
      type: object
      properties:
        memory_id: { type: integer, example: 11 }
        total_versions: { type: integer, example: 4 }
        timeline:
          type: array
          items: { type: object }
      required: [memory_id, total_versions, timeline]
paths:
...
  /authors/{author_id}/memories:
    get:
      tags: [Memories, Authors]
      summary: List memories for an author (RBAC + ownership enforced)
      security: [{ BearerAuth: [] }]
      parameters:
        - name: author_id
          in: path
          required: true
          schema: { type: integer, example: 1 }
      responses:
        "200":
          description: List of memories
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Memory" }
        "403":
          description: Forbidden (ownership)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "500":
          description: Server error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

    post:
      tags: [Memories, Authors]
      summary: Create a memory for an author (version 1; RBAC + ownership enforced)
      security: [{ BearerAuth: [] }]
      parameters:
        - name: author_id
          in: path
          required: true
          schema: { type: integer, example: 1 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MemoryCreateRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Memory" }
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: Forbidden (ownership)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "500":
          description: Server error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /memories/{memory_id}:
    get:
      tags: [Memories]
      summary: Get memory by id (RBAC + ownership enforced)
      security: [{ BearerAuth: [] }]
      parameters:
        - name: memory_id
          in: path
          required: true
          schema: { type: integer, example: 11 }
      responses:
        "200":
          description: Memory
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MemoryWithMeta" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: Forbidden (ownership)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

    put:
      tags: [Memories]
      summary: Update memory (creates a new version; RBAC + ownership enforced)
      security: [{ BearerAuth: [] }]
      parameters:
        - name: memory_id
          in: path
          required: true
          schema: { type: integer, example: 11 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MemoryUpdateRequest" }
      responses:
        "200":
          description: Updated (new version)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Memory" }
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: Forbidden (ownership)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /memories/{memory_id}/versions:
    get:
      tags: [Memories]
      summary: Get ledger events (versions) for a memory (RBAC + ownership enforced)
      security: [{ BearerAuth: [] }]
      parameters:
        - name: memory_id
          in: path
          required: true
          schema: { type: integer, example: 11 }
      responses:
        "200":
          description: Ledger events
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/LedgerEvent" }
        "403":
          description: Forbidden (ownership)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "500":
          description: Server error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /memories/{memory_id}/rollback/{version}:
    post:
      tags: [Memories]
      summary: Rollback memory to a previous version (creates a new version; RBAC + ownership enforced)
      security: [{ BearerAuth: [] }]
      parameters:
        - name: memory_id
          in: path
          required: true
          schema: { type: integer, example: 11 }
        - name: version
          in: path
          required: true
          schema: { type: integer, example: 1 }
      responses:
        "200":
          description: Rollback OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RollbackResponse" }
        "404":
          description: Version or memory not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: Forbidden (ownership)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "500":
          description: Server error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /memories/{memory_id}/timeline:
    get:
      tags: [Memories]
      summary: Get consolidated timeline for a memory (RBAC + ownership enforced)
      security: [{ BearerAuth: [] }]
      parameters:
        - name: memory_id
          in: path
          required: true
          schema: { type: integer, example: 11 }
      responses:
        "200":
          description: Timeline
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TimelineResponse" }
        "403":
          description: Forbidden (ownership)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "500":
          description: Server error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }